Added a basic Move class
