+++ src/main/java/janggi/master/Main.java
@@ -1,4 +1,4 @@
-package janngi_master;
+package janggi.master;
 
 import javax.swing.JFrame;
 import javax.swing.SwingUtilities;
