+++ src/main/java/janggi/master/core/Move.java
@@ -1,0 +1,7 @@
+package janggi.master.core;
+
+public class Move {
+    private Grid from, to;
+    private Camp camp;
+    private Piece piece;
+}
