+++ src/main/java/janggi/master/core/Piece.java
@@ -1,4 +1,4 @@
-package janggi_master.core;
+package janggi.master.core;
 
 public enum Piece {
     GENERAL, GUARD, HORSE, ELEPHANT, CHARIOT, CANNON, SOLDIER;
